requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2026.1
psycopg2-binary
Pillow
drf-spectacular==0.27.0
django-cors-headers==4.3.1
coverage==7.4.1
redis==5.2.1
